Preserve Unicode strings when passed to utility functions
This updates the string utility functions to return a Unicode
string when a Unicode string is passed in. In Python 2 this would
raise an error due to the use of str(v).

This is a change related to issue #374.

The tests also check for other types (int) to preserve existing
functionality.