Maybe propagates error information (#369)
Due to Any evaluating all options, and raising the exception
from the first encountered error, Maybe would discard the
expected error raised by the provided validator and only
raise a invalid value error (due to values not being None).

This commit changes the order of the validators in Maybe so
that the first evaluated error, and thus returned is that of
the provided validator.