Add error handling of incomparable values in Range
In Python3 some values that are comparable in Python2 are no longer
comparable. One instance of this is None, which in Python2 is always
less than any other object. In Python3, however, a TypeError is raised
if it is used in a comparison.

This commit handles said TypeError and issues a RangeInvalid exception
instead.