Feature: Support requiring anyOf a list of keys
This adds a new feature to Voluptuous, which is somewhat akin to what json-schema does with the special key `anyOf`.

`Schema({Required(Any('color', 'temperature', 'brightness')): str})` will validate that AT LEAST ONE of these three values is present. That doesn't preclude any individual validation on each of those fields to still apply.
That means that in the above example, if `color` is present, brightness doesn't need to be present. But if brightness is present, all other validations of brightness (like checking that its value is a number between 0 and 100) still apply.